Add files via upload
The object detector for detecting green caterpillars, other caterpillars and non-caterpillars created using YOLOv8 by Ultralytics with 100 epochs. The files consists of graphs, figures and examples of the training process and performance. The object detector can be used with "best.pt" for the best performing model and "last.pt" for the last iteration of the model.